Move class to maybe_pyfx
